chore(auth): Move AWSMobileClient operations to specific files for readability (#4015)
