+++ AWSAuthSDK/Sources/AWSMobileClient/Helpers/UserPoolHelper/AWSUserPoolOperationsHandler.swift
@@ -19,7 +19,7 @@
 import Foundation
 import AWSCognitoIdentityProvider
 
-protocol UserPoolAuthHelperlCallbacks {
+protocol UserPoolAuthHelperCallbacks {
 
     func getPasswordDetails(_ authenticationInput: AWSCognitoIdentityPasswordAuthenticationInput, passwordAuthenticationCompletionSource: AWSTaskCompletionSource<AWSCognitoIdentityPasswordAuthenticationDetails>)
     
@@ -61,7 +61,7 @@ AWSCognitoUserPoolInternalDelegate {
     internal var currentSignInHandlerCallback: ((SignInResult?, Error?) -> Void)?
     internal var currentConfirmSignInHandlerCallback: ((SignInResult?, Error?) -> Void)?
     
-    var authHelperDelegate: UserPoolAuthHelperlCallbacks?
+    var authHelperDelegate: UserPoolAuthHelperCallbacks?
     var customAuthHandler: AWSUserPoolCustomAuthHandler?
     internal static let sharedInstance: UserPoolOperationsHandler = UserPoolOperationsHandler()
 
@@ -114,7 +114,7 @@ AWSCognitoUserPoolInternalDelegate {
         return self
     }
     
-    internal func setAuthHelperDelegate(authHelperDelegate: UserPoolAuthHelperlCallbacks) {
+    internal func setAuthHelperDelegate(authHelperDelegate: UserPoolAuthHelperCallbacks) {
         self.authHelperDelegate = authHelperDelegate
     }
